requirements.txt
python-dotenv
requests
pillow
tmdbsimple
pytest
datetime